+++ httsleep/main.py
@@ -23,16 +23,6 @@ class HttSleeper(object):
                   single success condition dict.
     :param alarms: a list of error conditions, respresented by dicts, or a
                    single error condition dict.
-    :param status_code: deprecated. Shorthand for a success condition dependent on the
-                        response's status code.
-    :param json: Deprecated. Shorthand for a success condition dependent on the response's
-                 JSON payload.
-    :param jsonpath: Deprecated. Shorthand for a success condition dependent on the evaluation
-                     of a JSONPath expression.
-    :param text: Deprecated. Shorthand for a success condition dependent on the response's
-                 body payload.
-    :param callback: shorthand for a success condition dependent on a callback
-                     function that takes the response as an argument returning True.
     :param auth: a (username, password) tuple for HTTP authentication.
     :param headers: a dict of HTTP headers.  If specified, these will be merged with (and take
                     precedence over) any headers provided in the session.
@@ -53,16 +43,13 @@ class HttSleeper(object):
 
     """
     def __init__(self, url_or_request, until=None, alarms=None,
-                 status_code=None, json=None, jsonpath=None, text=None, callback=None,
                  auth=None, headers=None, session=DEFAULT_SESSION, verify=None,
                  polling_interval=DEFAULT_POLLING_INTERVAL,
                  max_retries=DEFAULT_MAX_RETRIES,
                  ignore_exceptions=None,
                  loglevel=logging.ERROR):
-        if status_code or json or jsonpath or text or callback:
-            msg = ("The shorthand kwargs status_code, json, jsonpath, text"
-                   " and callback are deprecated and will soon be removed.")
-            warnings.warn(msg, DeprecationWarning)
+        if not until:
+            raise ValueError("No success conditions provided!")
         if isinstance(url_or_request, string_types):
             self.request = requests.Request(
                 method='GET', url=url_or_request, auth=auth, headers=headers)
@@ -81,16 +68,6 @@ class HttSleeper(object):
             self.max_retries = int(max_retries)
         else:
             self.max_retries = None
-
-        if until is None and not (status_code or json or jsonpath or text or callback):
-            msg = ("No success conditions provided! Either the 'until' kwarg"
-                   " or one of the individual condition kwargs must be provided.")
-            raise ValueError(msg)
-        elif until is None:
-            until = []
-            condition = {'status_code': status_code, 'json': json,
-                         'jsonpath': jsonpath, 'text': text, 'callback': callback}
-            until.append({k: v for k, v in condition.items() if v})
 
         self.kwargs = {}
         if verify is not None:
@@ -203,8 +180,7 @@ class HttSleeper(object):
         return True
 
 
-def httsleep(url_or_request, until=None, alarms=None, status_code=None,
-             json=None, jsonpath=None, text=None, callback=None,
+def httsleep(url_or_request, until=None, alarms=None,
              auth=None, headers=None, session=DEFAULT_SESSION, verify=None,
              polling_interval=DEFAULT_POLLING_INTERVAL,
              max_retries=DEFAULT_MAX_RETRIES,
@@ -216,8 +192,7 @@ def httsleep(url_or_request, until=None, alarms=None, status_code=None,
     :return: :class:`requests.Response` object.
     """
     return HttSleeper(
-        url_or_request, until=until, alarms=alarms, status_code=status_code,
-        json=json, jsonpath=jsonpath, text=text, callback=callback,
+        url_or_request, until=until, alarms=alarms,
         auth=auth, headers=headers, session=session, verify=verify,
         polling_interval=polling_interval,
         max_retries=max_retries,